AI_image/1Prompt1Story/requirements.txt
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
opencv-python>=4.8.0 